Documentation is also important!
Documents `Marshale.ModelMagic`, and get doctests working as
planned. This allows us to know the docs (or at least the examples)
are up to date with the code. Much of the documentation is
extraneous, but this is a project to learn my way around Elixir
packaging, so hopefully that is fine.